Organize code into dedicated module files
This modular organization improves code maintainability by:

- Separating each component into its own file - Moving shared
functionality into base classes - Using __init__.py for clean public
APIs - Following single responsibility principle - Making components
independently testable
